Added operator overloads for type double.
The operators * *= / /= can use the type AT as first and/or second
argument.
This is now extended with the explicit type double!
Normally this should be possible implicitly if their is a ctor AT(double x).
For for some reason this does not work e.g. for AT=complex<double> or
others.
This way it is now posible to make calls like
Vec<..., complex<double>> x; auto y=x*3;

(This will be needed next for some special types for AT)