Use AT() as default initialzation value for Vector/Matrix initialized with INIT.
Before this commit the default value was 0. Hence, it is not changed if
AT is double or int (or complex<double>).
However, it gives the possibility to use other values for other type AT.
E.g. the new type AT=Expr uses not 0 but a independent symbolic variable.

If a Matrix is initialized with EYE the default value has not changed.
It is and stays for all AT 1.