Operator rename
Renamed the operator&= for setting independend symbolic variables to
operator^=. This is done to avoid confusion with the mat/vec operator&=
to subreference mat/vecs.
We need to use a operator since e.g. a setter function e.g. set(value)
cannot be implemented as a free function on mat/vec types.