Added new symbolic functions
Added the following functions to the AST:

atan2(y, x): calculates atan2(y, x)
min(a, b): calculates the minimum of a and b
max(a, b): calculates the maximum of a and b
heaviside(x): calculates 1 if x>0 else 0
condition(c, gt, le): if c>0 returns gt else le

Please note the these functions may have an undefined derivative or are even discontinuous
at specific points. The value at discontinuities is undefined (it may be the left or right limit).
The same holds for the derivative value of these points.
So take care when using these functions, especially when derivatives of these functions are needed.