Use dynamic memory also for a small size LRUCache and remove LRUCache from symbolicFunc
A symbolic function wrapper for a native function with a vector/matrix
return value should not itself use a cache to avoid a hierarchical cache
and because the cache should be introduced by the function which has
itself a high cost. This is not the case for the vector/matrix wrapper
functions.